Refactored signing tests to a different test file
