+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "npx hardhat test"
+    "test-configuration-tests": "npx hardhat test --dirname=configuration-tests",
+    "test-general-tests": "npx hardhat test --dirname=general-tests"
   },
   "author": "",
   "license": "ISC",
@@ -18,5 +19,8 @@
     "hardhat": "^2.6.2",
     "ts-node": "^10.4.0",
     "typescript": "^4.5.2"
+  },
+  "dependencies": {
+    "@toruslabs/starkware-crypto": "^1.0.0"
   }
 }
